src/requirements.txt
Django~=3.2
environs==9.5.0
psycopg2-binary==2.9.3
celery==5.2.7
redis==4.3.4
djangorestframework==3.13.1
drf-yasg==1.20.0
mysqlclient==2.1.1
pytest==7.1.2
django-constance==2.9.0
pydantic==1.9.1
django-filter~=22.1
gunicorn~=20.1
django-cors-headers~=3.13
django_celery_beat~=2.3
django-admin-inline-paginator~=0.3